+++ CHANGELOG.md
@@ -35,6 +35,10 @@
   });
   ```
 
+## Bug Fixes
+
+* Fixed an issue where output device test doesn't fail after providing a deviceId on browsers that doesn't support [setSinkId](https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/setSinkId);
+
 # 1.0.0-preview3 (June 17, 2020)
 
 Inital release and open source of project RTC Diagnostics SDK. This SDK provides developers with tools to diagnose potential problems before utilizing other Twilio SDKs such as Voice SDK.